数据库关闭问题

+++ OA/app/src/main/java/com/jxd/oa/activity/ExpenseAccountDetailActivity.java
@@ -63,9 +63,9 @@ public class ExpenseAccountDetailActivity extends AbstractActivity {
             auditUser_tv.setText(expenseAccount.getAuditUser().getName());
         }
         auditRemark_tv.setText(expenseAccount.getAuditRemark());
-        if (expenseAccount.getAuditStatus() == Const.STATUS_AUDIT_REFUSE.getValue()) {
+        if (expenseAccount.getAuditStatus() ==(Integer)  Const.STATUS_AUDIT_REFUSE.getValue()) {
             auditStatus_tv.setTextColor(mContext.getResources().getColor(R.color.color_red));
-        } else if (expenseAccount.getAuditStatus() == Const.STATUS_AUDIT_PASS.getValue()) {
+        } else if (expenseAccount.getAuditStatus() ==(Integer)  Const.STATUS_AUDIT_PASS.getValue()) {
             auditStatus_tv.setTextColor(mContext.getResources().getColor(R.color.color_being));
         } else {
             auditStatus_tv.setTextColor(mContext.getResources().getColor(R.color.color_black_font));
