+++ OA/app/src/main/java/com/jxd/oa/adapter/LeaveApplicationAdapter.java
@@ -38,9 +38,9 @@ public class LeaveApplicationAdapter extends AbstractAdapter<LeaveApplication> {
         }
         viewHolder.title_tv.setText(getItem(position).getLeaveReason());
         viewHolder.date_tv.setText(getItem(position).getStartDate() + "至" + getItem(position).getEndDate());
-        if (getItem(position).getAuditStatus() == Const.STATUS_AUDIT_REFUSE.getValue()) {
+        if (getItem(position).getAuditStatus() == (Integer) Const.STATUS_AUDIT_REFUSE.getValue()) {
             viewHolder.status_tv.setTextColor(getContext().getResources().getColor(R.color.color_red));
-        } else if (getItem(position).getAuditStatus() == Const.STATUS_AUDIT_PASS.getValue()) {
+        } else if (getItem(position).getAuditStatus() ==(Integer)  Const.STATUS_AUDIT_PASS.getValue()) {
             viewHolder.status_tv.setTextColor(getContext().getResources().getColor(R.color.color_being));
         } else {
             viewHolder.status_tv.setTextColor(getContext().getResources().getColor(R.color.color_black_font));
